Troubleshooting device
The example file for the FINFET device, in which no device layers are
shared, worked but this was not the case for the MOSFET, in which device
layers are shared or copied. It was found that using the different copy
methods, either layer by layer copying or direct device copying, led to
the same, undesired result. In particular, the features of a layer which
was stacked earlier were incremented to the same level as the most
recently stacked (top) device in the schematic.