improved __init__ method of Layer
All arguments passed to the initialization are assigned attributes at
the beginning and are referred to later by the attributes. The
initialization is sequenced in this attribute assignment, default
handling of attributes (which is less efficient than a set of if-else
statements that would only assign a non-default value if non-default,
but is used for clarity), and feature definition.