path clipping and object copying bugs
A debugging option to enable/disable pathclipping shows for the example
cases that the device and schematic creation work as intended but it is
further required

One design problem seems to be that the attributes of objects nested
within other objects are changed when those higher level objects run
methods. It may be better design principle to have each object store
the relevant position information and then only on writing the highest
level object (schematic) have them add. This is possible since all the
position changes add independently (the phase shift and replication
of a feature in a layer, the vertical shift due to layer in device
placement, and the vertical and horizontal shift due to device in
schematic placement)