worked examples with changes
Changes as suggested in 545f36985fcea0a7110ab8f9523a04baff07560e.

There were problems in that copy methods of higher level objects did not
copy all of the objects they contained (layer did not copy feature), so
the conformal layer and layer shared the same feature.

The stack_heights routine was changed to not be a cumulative sum. Instead it stores
the most recent stack height, and then increases its repetition for the
number of same plane layers in the next stack operation.

At this point the code is still messy since
- the coordinates of the feature are changed by conformal_layer but nothing
else does
- there isn't a consistent approach to the features and most don't have
all the required methods
- the copy methods are awkward. One example, to copy the device the
__init__ has to take as inputs those things which define the device (or set its
attributes, like is done in conformal layer). But in making a device only
the default values are used.