major revisions to class structure and methods
the Feature class has been simplified to have, for polygon features, a
superclass PolygonFeature which defines a drawing method for polygon.
the subclasses now only define a set of coordinates for the vertices of
a polygon.

the main advantage in the change of the layer, device, and schematic
classes is recursive definitions of the place method (not actually
recursive since it calls the place method of lower classes rather than
itself, but similar in idea). Because of the recursive definition it is
possible to draw any object which has the place method which is useful for debugging.

some features such as deep copying and clipping were removed in the
refactor but will be added back in.

examples except for the most recent test example are broken and need to
be made compatible with the changes