copy methods reimplemented for device, layer
features no longer need to be deep copied since the layer only has a
place method using a single feature. this is somewhat inconsistent in
design since the layer isn't a list of features, but the device is
a list of layers. the difference is the layer is now defined as an
infinitely periodic structure in the x-direction and finite in the
y-direction, so it can't be a list of features until the device width
is specified. in order to allow layers to be defined independently of
devices the place method, which is only called from a containing device
method (or with a user supplied width), is the only one which places
features

also copy method implemented for schematic, but this is for completeness
only and a rare use case