shapes example file and convenience methods update
the shapes example file update required fixes to the place method in the
layer.

added get height and get width methods to features since feature
dimensions factor into placement for a given device width.

clipping still not supported, as of now the feature is simply not drawn
if any part of it goes over the boundary. there may be cases where
clipping is desired for features, e.g., to clip half at the left
boundary and half at the right boundary for a symmetric feature (which
would also require, or be best if implemented with, negative phase
shifts)